+++ tests/testthat/test-agg_helpers.R
@@ -1,4 +1,4 @@
-test_that("df_to_aggregation_map() works as expected", {
+test_that("agg_table_to_agg_map() works as expected", {
   bands <- tibble::tribble(~few_col, ~many_col, 
                            "The Beatles", "John", 
                            "The Beatles", "Paul", 
@@ -12,7 +12,7 @@ test_that("df_to_aggregation_map() works as expected", {
                            "Rolling Stones", "Ronnie",
                            "Rolling Stones", "Bill",
                            "Rolling Stones", "Charlie")
-  res <- df_to_aggregation_map(bands, 
+  res <- agg_table_to_agg_map(bands, 
                                few_colname = "few_col",
                                many_colname = "many_col")
   expect_equal(res, list(`Rolling Stones` = c("Mick", "Keith", "Ronnie", "Bill", "Charlie"),
@@ -20,10 +20,10 @@ test_that("df_to_aggregation_map() works as expected", {
 })
 
 
-test_that("aggregation_map_to_df() works as expected", {
+test_that("agg_map_to_agg_table() works as expected", {
   agg_map <- list(`The Beatles` = c("John", "Paul", "George", "Ringo"), 
                   `Rolling Stones` = c("Mick", "Keith", "Ronnie", "Bill", "Charlie"))
-  res <- aggregation_map_to_df(agg_map, few_colname = "band", many_colname = "member")
+  res <- agg_map_to_agg_table(agg_map, few_colname = "band", many_colname = "member")
   expect_equal(res, 
                tibble::tribble(~band, ~member, 
                                "The Beatles", "John",